requirements.txt
Flask
gunicorn
flask_sqlalchemy
sqlalchemy
pymysql
mysql-connector-python